Separate visual and app layer in staff members
