+++ src/staff/components/StaffListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./StaffListPage";
 export * from "./StaffListPage";
+export * from "./filters";
