--- src/staff/types.ts
@@ -1,6 +1,0 @@
-import { FilterOpts } from "@saleor/types";
-import { StaffMemberStatus } from "@saleor/types/globalTypes";
-
-export interface StaffListFilterOpts {
-  status: FilterOpts<StaffMemberStatus>;
-}
