+++ src/staff/views/StaffList/StaffList.tsx
@@ -46,7 +46,7 @@ import {
   saveFilterTab,
   getFilterQueryParam,
   getFilterOpts
-} from "./filter";
+} from "./filters";
 import { getSortQueryVariables } from "./sort";
 
 interface StaffListProps {
