--- src/staff/views/StaffList/messages.ts
@@ -1,18 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  active: {
-    defaultMessage: "Active",
-    description: "staff member's account"
-  },
-  deactivated: {
-    defaultMessage: "Deactivated",
-    description: "staff member's account"
-  },
-  status: {
-    defaultMessage: "Status",
-    description: "staff member's account"
-  }
-});
-
-export default messages;
